+++ src/index.js
@@ -101,6 +101,11 @@ export { FormError } from "./components/FormError/FormError";
 export { FormLabel } from "./components/FormLabel/FormLabel";
 
 /**
+ * Form RadioButton
+ */
+export { FormRadioButton } from "./components/FormRadioButton/FormRadioButton";
+
+/*
  * Check box
  */
 export { CheckBoxes } from "./components/CheckBoxes/CheckBoxes";