+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Text Input Field
 - FormError Component
 - Form Label Component
+- Form RadioButton Component
 - Checkboxes Component
 
 ### Changed